Fixed kernel stub argument marshalling code to ensure proper alignment
Previously, shamrock was generating load instructions in the kernel stub with
a strict type alignment, which was not being followed by the argument
marshalling code.

This resulted in a NEON vld1 instruction failing in the kernel stub,
and trashing a base register which was not 16 byte aligned.

Now the marshalling code calculates the proper alignement for each argument,
and is based on a buffer aligned to double16 to begin with.

With this patch, all the vector sub-tests of the Khronos basic_parameter_types
test now pass.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>