Added -ffake-address-space-map option to generate LLVM IR address spaces
Previously, shamrock was not able to get address space attributes from
LLVM's getAddressSpace() API for function (kernel) arguments, as they
didn't exist in the generated LLVM IR.

Now, we use a clang option to force address space mapping, causing
addrspace(<n>) attrs to be inserted.

See the discussion here:
https://www.mail-archive.com/cfe-commits@cs.uiuc.edu/msg76370.html

This was a necessary step to get many of the Khronos basic tests involving
local variables and parameters to not crash.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>