Add support for passing union and struct arguments by value to kernels.
Previously, shamrock was able to accept pointers to aggregate types.
However, the new v 1.2 api test 'test_get_kernel_arg_info' creates a few
kernels with struct and union arguments passed by value, causing that
test to fail.

Unfortunately, LLVM support for Unions has been officially removed, so
Unions in IR are reported as Structs.  Also, we note that in the test case,
LLVM Structs are reported as Arrays.

Fortunately, LLVM APIs are correctly reporting the target size and alignment
for those types, which seems to be sufficient for recognizing, setting
and passing arguments correctly.

So, a new Kind::StructType is added, and the target size and alignment are
stored for this case.  This seems sufficient to handle the LLVM IR 'union_type'.

Array types are treated as vectors, which sets a vector dimension, and
element size.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>