Implement OpenCL v1.2 API clGetKernelArgInfo()
Kernel argument metadata is generated when adding the clang switch:
  -cl-kernel-arg-info

In this case, it is possible to walk the kernel metadata for the compiled
module, and get information about each kernel's arguments.

Upon calling Kernel::addFunction(), typically from createKernel(),
this data is parsed and stored for later retrieval by the clGetKernelArgInfo().

Much of the parsing code was derived from review of the pocl and mesa/clover
clGetKernelArgInfo() implementations.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>