Implement "Embedded Header" feature of the v1.2 Separate Compiling and Linking
v1.2 adds the ability to load header file source into program objects,
then pass those program objects into clCompileProgram().

Without the ability to compile the program into LLVM IR, or to have
a functional VFS in clang (?), this forces us to create temporary
actual header files on the real filesystem for clang to locate.

This commit causes the header source to be written into a temporary
/tmp/.shamrock directory, and then adds the "-I /tmp/.shamrock" include
path to the compile options.

This was validated using Khronos tests:
% test_compiler simple_embedded_header_compile
% test_compiler simple_embedded_header_link
% test_compiler execute_after_embedded_header_link
% test_compiler multiple_embedded_headers

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>