Moved build notifications out of the program object to the API level
Per the OCL 1.2 spec, build (compile, link and build) APIs must call the
provided callback even if the builds fail.

Previously, this was only being done when the builds succeeded, and also
was being done once per device (which the spec does not require).

This has been changed so that notifications now occur for *any* build result.

This was validated using the following test_compiler Khronos test cases:

	% test_compiler simple_compile_with_callback
	% test_compiler simple_link_with_callback
	% test_compiler execute_after_simple_compile_and_link_with_callbacks
	% test_compiler simple_library_with_callback

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>