Implemented clEnqueueFillBuffer() v1.2 API.
This implementation is quite simple, and avoids creation of special
OpenCL kernels in favor of using a GNU host builtin, which it is expected
would be sufficiently optimized.

A possible optimization worth exploring would be to use OpenMP on the
host to dispatch the pattern filling across the host cores.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>