relaxed comparison of Args objects in Kenrel objects
Previously, comparison of Args objects between Kernel objects returned
false if the Args this pointers were not the same.

For two different devices with two separate kernels, it is still possible the
Args are the same, if their data members are the same.

So, replaced the != operator with an comparison method, which avoids
inheriting the C++ != operator which was comparing this pointers.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>