clCreateSubDevices():  Return CL_DEVICE_PARTITION_FAILED if numCPUs() == 1
If a sub-device cannot be further sub-divided, the v1.2 spec allows
the implementation to return CL_DEVICE_PARTITION_FAILED, or CL_INVALID_VALUE
depending on one's interpretation.

The Khronos v1.2 device_partition test fails in either case (which appears
to be a bug).

The Khronos v2.0 device_partition test was modified to accept the
return value of CL_DEVICE_PARTITION_FAILED, which is a logical fix.

This patch allows a similarly modified (fixed) Khronos v1.2 test to
pass the test subroutine code which recursively subdivides devices.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>