Sub-devices inherit the parent's kernel and program device dependent structs.
This implements an aspect of the clCreateSubDevices() API, per the v1.2 spec:

    "A program binary (compiled binary, library binary or executable binary)
    built for a parent device can be used by all its sub-devices.
    If a program binary has not been built for a sub-device, the
    program binary associated with the parent device will be used."

Previously, each program or kernel object was created for a device.  If
a device was paritioned into sub-devices, the sub-devices may not have
had an associated program/kernel object (unless one was explicitly built).

Now, when shamrock queries a device for a kernel or program, it will search
the device hierarchy and look for a parent device for which a program/kernel
object may have been created.

This patch enables the Khronos v1.2 test_device_partition to PASS.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>