Implement clEnqueueMigrateMemObjects() v1.2 API
This implementation passes the Khronos test_buffer_migrate test of the
buffers test.

However, a real test requires devices of different memory affinities,
between which to migrate the buffers, which
in the case of a CPUDevice only implementation on heterogenous cores,
we don't currently have.

Nevertheless, this patch implements the API, and the backend has a
nominal implementation which will pre-allocate device buffers when called.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>