Update Kernel::reqdWorkGroupSize() to search all meta node operands.
With an update of LLVM to 3.6+, the reqd_work_group_size attribute is no
longer sitting in the first operand position of the function's meta data
node. It now can be anywhere.

This patch iterates through all the metanode operands searching for the
reqd_work_group_size attribute.

This enables the Khronos test:
	% test_api kernel_required_group_size
to PASS.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>