Explicitly return CL_INVALID_VALUE for CL_KERNEL_GLOBAL_WORK_SIZE case
Per the v1.2 spec, for clGetKernelWorkGroupInfo(), CL_KERNEL_GLOBAL_WORK_SIZE:

"If device is not a custom device or kernel is not a built-in kernel,
clGetKernelArgInfo returns the error CL_INVALID_VALUE."

Since the CPU is not a custom device, and has no built-in
kernels, then the return value of CL_INVALID_VALUE is correct.

This reverts commit 668bdd17cadc838a8bc165a0f222e31c7b6777a7, but as
a result may cause an error in piglit testing, which that commit
was meant to resolve.

Signed-off-by: Gil Pitney <gil.pitney@linaro.org>